rework import strat. top-down: game->game_1->game_setup
