+++ src/game_setup.py
@@ -8,6 +8,7 @@ import discord
 from .configuration import Configuration
 from .db import get_games_w_status, GameStatus, get_game_from_id, update_db_obj
 from .db import GameStatus, Game
+from .game_1 import finish_game_1
 
 
 class StatusSelect(discord.ui.Select):
@@ -246,7 +247,8 @@ class ConfirmationView(discord.ui.View):
         await super().on_timeout()
 
 
-async def evaluate_game(interaction: discord.Interaction, config: Configuration):
+async def evaluate_game(interaction: discord.Interaction, config: Configuration, bot):
+    print(type(bot))
     games = await get_games_w_status(config, [GameStatus.STOPPED])
     select_view = GenGameSelectView(config, games)
     await interaction.response.send_message(
@@ -274,7 +276,7 @@ async def evaluate_game(interaction: discord.Interaction, config: Configuration)
     # Idee 1
     match game.name:
         case "Fast and hungry, task hunt":
-            await finish_game_1
+            await finish_game_1(config, game, bot)
         case _:
             config.watcher.logger.error(
                 f"Game with ID {game.id} has an unknown name: {game.name}."
